python-worker/requirements.txt
# AI照片生成Worker依赖
redis>=4.0.0
torch>=1.8.0
torchvision>=0.9.0
pillow>=8.0.0
numpy>=1.19.0
tqdm>=4.60.0
requests>=2.25.0
matplotlib>=3.4.0
scikit-image>=0.18.0
scipy>=1.6.0
transformers>=4.5.0 